% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df}
\alias{df}
\title{Touring Plans Wait Time Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 31059 rows and 195 columns.}
\source{
\href{https://touringplans.com/walt-disney-world/crowd-calendar#DataSets}{Touring Plans Datasets}
}
\usage{
data(df)
}
\description{
Touring Plans Wait Time Data
}
\examples{
data(df)
}
\references{
Disney World Ride Wait Time Datasets, TouringPlans.com, June 2018,
https://www.touringplans.com/walt-disney-world/crowd-calendar/#DataSets, Accessed 28 June 2018.
}
\keyword{datasets}
