% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn_in_challenge.R
\name{turn_in_challenge}
\alias{turn_in_challenge}
\title{Turn in challenge results, launch final survey}
\usage{
turn_in_challenge()
}
\description{
This will open a browser with a final survey to submit your results.
If you do not have pop-ups enabled, you will get an error message -
click "Try Again" and the survey should load.
}
